requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.25.2
plotly==5.17.0
dash==2.13.0
scikit-learn==1.3.0
statsmodels==0.14.0
pytest==7.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
flask-wtf==1.1.1
